python/src/plugins/spl_token/pyproject.toml
[tool.poetry]
name = "goat-sdk-plugin-spl-token"
version = "0.1.4"
description = "Goat plugin for spl_token"
authors = [
    "Your Name <your_email@example.com>",
]
readme = "README.md"
keywords = [
    "goat",
    "sdk",
    "agents",
    "ai",
    "spl_token",
]
homepage = "https://ohmygoat.dev/"
repository = "https://github.com/goat-sdk/goat"
packages = [
    { include = "goat_plugins/spl_token" },
]

[tool.poetry.dependencies]
python = "^3.10"
goat-sdk = "*"
goat-sdk-wallet-solana = "*"
solders = ">=0.18.0"

[tool.poetry.dependencies.solana]
version = ">=0.30.2"
extras = [
    "spl",
]

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.6"

[tool.poetry.group.dev.dependencies.goat-sdk]
path = "../../goat-sdk"
develop = true

[tool.poetry.group.dev.dependencies.goat-sdk-wallet-solana]
path = "../../wallets/solana"
develop = true

[tool.poetry.group.dev.dependencies.solana]
version = ">=0.30.2"
extras = [
    "spl",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/goat-sdk/goat/issues"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "src"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120
target-version = "py312"

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
